decimal

+++ Abstractions/ExchangeClients.Abstractions/IExchangeClient.cs
@@ -2,5 +2,5 @@ namespace Exchanges.Abstractions;
 
 public interface IExchangeClient
 {
-    Task<int> GetExchangePairPriceAsync(string pairName, CancellationToken ct);
+    Task<decimal> GetExchangePairPriceAsync(string pairName, CancellationToken ct);
 }
